FFT error
There is an error is the FFT example. The peak frequency is found to be 0.2 which is correct, however the plot insert shows the peak frequency at 0.25. This is because the pos_mask is not applied to the power array. In the worked out example code the line "plt.plot(freqs[:8], power[:8])" should be edited to "plt.plot(freqs[:7], power[pos_mask][:7])". This applies the mask to power as well as frequency and shows the correct result in the figure insert.